* Split RoutesMapper into separate responsibilities to reduce coupling between objects.
